Combine Process Uploads and Upload Samples * Removes the need for intermediate status * Removes the need for get samples from db function
