+++ scripts/process_uploads.py
@@ -5,32 +5,77 @@ into the MAIN and ARCHIVE buckets using the batch_move_files function.
 
 Assumes that the csv file has been processed in a previous step. 
 """
-
+import csv
+import io
 import os
+import json
 from os.path import join
 from typing import List, Optional, Tuple
 import click
 import hailtop.batch as hb
+from google.cloud import storage
 
 # Import SampleAPI
 from sample_metadata.api.sample_api import SampleApi
 from joint_calling.upload_processor import batch_move_files, SampleGroup
 
 
-# Reading CSV pulled out into separate function.
+def get_csv(bucket_name, prefix) -> Tuple[csv.DictReader, str]:
+    """Pull relevant data from csv file in a given bucket
+    ==========
+    Only one .csv file exists in the upload bucket and previous batch directory.
+    """
+    client = storage.Client()
+    gcs_bucket = client.get_bucket(bucket_name)
+
+    all_blobs = list(client.list_blobs(bucket_name, prefix=prefix))
+    csv_path = next(filter(lambda blob: blob.name.endswith('.csv'), all_blobs)).name
+
+    csv_as_text = gcs_bucket.get_blob(csv_path).download_as_text()
+    csv_reader = csv.DictReader(io.StringIO(csv_as_text))
+
+    return csv_reader, csv_path
 
 
-def get_samples_from_db(proj) -> List[str]:
-    """ Pulls list of samples to be moved from SampleMetadata DB """
+def create_analysis(csv_dict_reader, proj):
+    """New analysis objects created for the gvcf and cram for each sample
+    Assumptions:
+    SampleSequence object created previously.
+    We are exclusively processing crams and gvcfs.
+    Samples with gvcfs have crams, and samples without gvcfs don't.
+    """
 
+    # Pull a list containing the sample ID's that don't have gvcfs
     sapi = SampleApi()
-
-    samples_internal_ids = sapi.get_samplesequencing_with_status(
-        proj, 'gvcf', 'Ready'
+    previous_samples_internal = sapi.get_samples_with_gvcfs(proj)  # TODO: Implement
+    previous_samples_external = sapi.get_external_ids(
+        previous_samples_internal, proj
     )  # TODO: Implement
-    samples_external_ids = sapi.get_external_ids(samples_internal_ids, proj)
 
-    return samples_external_ids
+    samples: List[str] = []
+    sample_metadata = []
+
+    for sample_dict in csv_dict_reader:
+        samples.append(sample_dict['sample.sample_name'])
+        sample_metadata.append(sample_dict)
+
+    # Determine the samples in the latest upload.
+    latest_upload_external = samples - previous_samples_external
+    latest_upload_internal = sapi.get_internal_ids(latest_upload_external, proj)
+
+    sample_meta_map = {d['sample.sample_name']: d for d in sample_metadata}
+    for sample in latest_upload_internal:
+        metadata = sample_meta_map[sample]
+        metadata_json = json.dumps(list(metadata)[0], indent=2)
+
+        sapi.create_analysis_object(sample, 'gvcf')
+        sapi.create_analysis_object(sample, 'cram')
+
+        sapi.update_metadata(sample, metadata_json)  # TODO: IMPLEMENT.
+
+        # sapi.update_sequencing_status(sample, 'Ready')  # TODO: IMPLEMENT.
+
+    return latest_upload_external
 
 
 def generate_file_list(
@@ -78,7 +123,7 @@ def update_status(sample_group: SampleGroup):
     """ Updates the status of a SampleSequence """
     sapi = SampleApi()
     sapi.update_sequencing_status(
-        sample_group.sample_id_external, 'Uploaded'
+        sample_group.sample_id_external, 'Upload Successful'
     )  # TO IMPLEMENT: API CALL
 
 
@@ -125,18 +170,24 @@ def run_processor(
     main_bucket = f'cpg-{project}-main'
     main_prefix = join('gvcf', batch_path)
     main_path = join(main_bucket, main_prefix)
+    metadata_bucket = f'cpg-{project}-main-metadata'
     archive_path = join(f'cpg-{project}-archive', 'cram', batch_path)
 
     docker_image = os.environ.get('DRIVER_IMAGE')
     key = os.environ.get('GSA_KEY')
 
     # Determine the analysis results (i.e. list of gvcfs and crams) to be moved
-    samples: List[str] = []  # List of external sample IDs
+    samples_external_ids: List[str] = []  # List of external sample IDs
     main_files: List[SampleGroup] = []
     archive_files: List[SampleGroup] = []
 
-    samples = get_samples_from_db(project)
-    main_files, archive_files = generate_file_list(samples)
+    #  Get the CSV file.
+    csv_reader, csv_path = get_csv(upload_bucket, upload_prefix)
+
+    samples_external_ids = create_analysis(csv_reader, project)
+
+    # samples = get_samples_from_db(project)
+    main_files, archive_files = generate_file_list(samples_external_ids)
 
     service_backend = hb.ServiceBackend(
         billing_project=project,
@@ -145,6 +196,7 @@ def run_processor(
 
     batch = hb.Batch(name='Process files', backend=service_backend)
 
+    main_jobs = []
     for sample_group in main_files:
         # Moving the files to the main bucket
         sample_group_main_jobs = batch_move_files(
@@ -171,7 +223,9 @@ def run_processor(
         )
         status_job.call(update_status, sample_group)
         status_job.depends_on(validate_job)
+        main_jobs.append(status_job)
 
+    archive_jobs = []
     for sample_group in archive_files:
         # Moving the files to the archive bucket
         sample_group_archive_jobs = batch_move_files(
@@ -196,6 +250,14 @@ def run_processor(
         )
         status_job.call(update_status, sample_group)
         status_job.depends_on(validate_job)
+        archive_jobs.append(status_job)
+
+    # Move the csv file to the metadata bucket, after the gVCFs and CRAMs have been
+    # moved to the main and archive buckets.
+    csv_job = setup_job(batch, f'Move {csv_path}', docker_image)
+    csv_job.command(f'gsutil mv gs://{csv_path} gs://{metadata_bucket}/{batch_path}/')
+    csv_job.depends_on(*main_jobs)
+    csv_job.depends_on(*archive_jobs)
 
     batch.run()
